♻️ remove duplicated code
