+++ src/functions/searches/attach-to-one-explorer-search/attach-to-one-explorer-search.ts
@@ -53,7 +53,7 @@ import { toDataExplorerEntry } from '~/models/search/to-data-explorer-entry';
 import { ID, Percentage, toNumericID } from '~/value-objects';
 import { APIContext, debounceWithBackoffWhile } from '../../utils';
 import { attachSearch } from '../attach-search';
-import { createRequiredSearchFilterObservable } from '../helpers/create-required-search-filter-observable';
+import { createRequiredSearchFilterObservable, DateRange } from '../helpers/create-required-search-filter-observable';
 import { makeSubscribeToOneRawSearch } from '../subscribe-to-one-raw-search';
 import { RawRequestExplorerSearchEntriesWithinRangeMessageSent } from '~/models';
 import {
@@ -64,7 +64,8 @@ import {
 	RequiredSearchFilter,
 	SEARCH_FILTER_PREFIX,
 } from '../subscribe-to-one-search/helpers';
-import { createInitialSearchFilter } from '~/functions/searches/helpers/create-initial-search-filter';
+import { createInitialSearchFilter } from '~/functions/searches/helpers/attach-search';
+import { getPreviewDateRange, DEBOUNCE_OPTIONS } from '../helpers/attach-search';
 
 export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 	const subscribeToOneRawSearch = makeSubscribeToOneRawSearch(context);
@@ -134,27 +135,12 @@ export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 		);
 		const rendererType = searchAttachMsg.data.RendererMod;
 
-		type DateRange = { start: Date; end: Date };
-		const previewDateRange: DateRange = await (async (): Promise<DateRange> => {
-			// Not in preview mode, so return the initial filter date range, whatever, it won't be used
-			if (initialFilter.dateRange !== 'preview') return initialFilter.dateRange;
-
-			// In preview mode, so we need to request search details and use the timerange that we get back
-			const detailsP = firstValueFrom(
-				searchMessages$.pipe(filter(filterMessageByCommand(SearchMessageCommands.RequestDetails))),
-			);
-			const requestDetailsMsg: RawRequestSearchDetailsMessageSent = {
-				type: searchTypeID,
-				data: { ID: SearchMessageCommands.RequestDetails },
-			};
-			rawSubscription.send(requestDetailsMsg);
-			const details = await detailsP;
-
-			return {
-				start: new Date(details.data.SearchInfo.StartRange),
-				end: new Date(details.data.SearchInfo.EndRange),
-			};
-		})();
+		const previewDateRange: DateRange = await getPreviewDateRange({
+			initialFilter,
+			searchTypeID,
+			searchMessages$,
+			rawSubscription,
+		});
 
 		const close = async (): Promise<void> => {
 			if (closed) return undefined;
@@ -268,15 +254,6 @@ export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 			const detailsResults = await Promise.all([detailsP, detailsMsgP]);
 			const detailsMsg = detailsResults[1];
 
-			// Dynamic duration for debounce a after each event, starting from 1s and increasing 500ms after each event,
-			// never surpass 4s, reset to 1s if the request is finished
-			const debounceOptions = {
-				initialDueTime: 1000,
-				step: 500,
-				maxDueTime: 4000,
-				predicate: (isFinished: boolean) => !isFinished, // increase backoff while isFinished is false
-			};
-
 			// Keep sending requests for search details until Finished is true
 			pollingSubs.add(
 				rawSearchDetails$
@@ -288,7 +265,7 @@ export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 						map(details => (details ? details.data.Finished : false)),
 
 						// Add dynamic debounce after each message
-						debounceWithBackoffWhile(debounceOptions),
+						debounceWithBackoffWhile(DEBOUNCE_OPTIONS),
 
 						// Filter out finished events
 						rxjsFilter(isFinished => isFinished === false),
@@ -321,7 +298,7 @@ export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 						map(entries => (entries ? entries.finished : false)),
 
 						// Add dynamic debounce after each message
-						debounceWithBackoffWhile(debounceOptions),
+						debounceWithBackoffWhile(DEBOUNCE_OPTIONS),
 
 						// Filter out finished events
 						rxjsFilter(isFinished => isFinished === false),
@@ -350,7 +327,7 @@ export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 						map(stats => stats.data.Finished ?? false),
 
 						// Add dynamic debounce after each message
-						debounceWithBackoffWhile(debounceOptions),
+						debounceWithBackoffWhile(DEBOUNCE_OPTIONS),
 
 						// Filter out finished events
 						rxjsFilter(isFinished => isFinished === false),
@@ -388,7 +365,7 @@ export const makeAttachToOneExplorerSearch = (context: APIContext) => {
 						map(stats => stats.data.Finished ?? false),
 
 						// Add dynamic debounce after each message
-						debounceWithBackoffWhile(debounceOptions),
+						debounceWithBackoffWhile(DEBOUNCE_OPTIONS),
 
 						// Filter out finished events
 						shareReplay({ bufferSize: 1, refCount: true }),
